Restructure the File errors in controller (#23801)
